training/requirements.txt
tensorflow
scikit-learn
google-cloud-storage
pyyaml